boiler/requirements.txt
Flask==0.9
Werkzeug==0.8.3
Jinja2==2.6
Flask-Assets==0.7
webassets==0.7
Flask-WTF==0.8
WTForms==1.0.2
Flask-Script==0.5.1
Flask-Mail==0.7.3
blinker==1.2
Flask-Cache==0.9.0
Fabric==1.4.3
ssh==1.7.14
pycrypto==2.6
python-memcached==1.48
redis==2.7.1
nose==1.1.2
Flask-DebugToolbar==0.7.1
celery-with-redis==3.0
simplejson==2.6.2
Flask-Login==0.1.3
fuzzywuzzy==0.1
Flask-Uploads==0.1.3
pymongo==2.7.1
tweepy
boto
sendgrid==0.1.4
twilio
requests
google-api-python-client==1.4.0
googlemaps==2.2
geoip2==2.2.0
requests-futures==0.9.5